+++ index.ts
@@ -1,2 +1,2 @@
 export { createRouter } from './src/cloudflare-router'
-export { BasicAuthMethod, IPAuthMethod, AuthMethods, Route, Config, Routes, DEFAULT_CONFIG } from './src/config'
+export { BasicAuthMethod, IPAuthMethod, AuthMethods, Route, Config, Routes } from './src/config'